fix warnings
Former-commit-id: b1049f7cf998bf35965b1140921babeb4a746040 [formerly 60f48ff0def4c63fe12250dc9f5ae6d83a1a4801] [formerly 625ce808f11763c1e28e97dd02b8ec2287b1c7ac [formerly 82f6087f625e5d462eebb84688cbbd6e0ae7cc03]]
Former-commit-id: 2dec5d8f4e0f12e2f807d4e0c88b72c5b2dfa1fe [formerly 970fc412792af123fae93ccfe7bfbbfefb05ab34]
Former-commit-id: adf9df7747329c4cf521309f92c7cad8c35adbb7

+++ Frame/Server/GameServer/GameNetServerPlugin/AFCGameNetServerModule.cpp
@@ -645,7 +645,7 @@ int AFCGameNetServerModule::OnCommonDataTableEvent(const AFGUID& self, const DAT
                         AFINetServerModule::RecordToPBRecord(xRowDataList, nRow, nCol, *pAddData);
                     }
 
-                    for(int i = 0; i < valueBroadCaseList.GetCount(); i++)
+                    for(size_t i = 0; i < valueBroadCaseList.GetCount(); i++)
                     {
                         AFGUID identOther = valueBroadCaseList.Object(i);
